Initial GPS data translations
Begin work on GPS data translations

+++ src/lidar_math/m_atoxyz.cpp
@@ -3,9 +3,9 @@
 point_t atoxyz(double a1, double a2, double distance) {
 	point_t p;
 
-	p.x = distance * cos(a1) * sin(a2);
-	p.y = distance * cos(a1) * cos(a2);
-	p.z = distance * sin(a1);
+	p.x = (float)(distance * cos(a1) * sin(a2));
+	p.y = (float)(distance * cos(a1) * cos(a2));
+	p.z = (float)(distance * sin(a1));
 
 	return p;
 }
